tests: toyserver: move to subdirectory
